refactor(runtime): isolate Box volume storage ownership
